用户加入收费活动的服务端实现，生成订单之前的信息验证，订单支付之后的操作逻辑

+++ src/main/java/com/labelwall/activity/dao/ActivityAccountMapper.java
@@ -23,5 +23,5 @@ public interface ActivityAccountMapper {
                                @Param("jindouCount") Integer jindouCount);
 
     int updateAccountSubJindouNum(@Param("accountId") Long accountId,
-                                  @Param("jindouCount") BigDecimal jindouCount);
+                                  @Param("jindouCount") Integer jindouCount);
 }
